+++ src/components/Buttons/Button/@types/index.ts
@@ -2,6 +2,7 @@ import type { MouseEventHandler, ReactNode } from "react";
 
 export type ButtonProps = {
   type?: "button" | "submit";
+  layout?: "cancel" | "remove" | "apply";
   className?: string;
   loadingClassName?: string;
   children?: ReactNode;