+++ src/components/Buttons/Button/index.tsx
@@ -1,13 +1,14 @@
 import classNames from "classnames";
 import styles from "../Buttons.module.css";
-import { ButtonProps } from "./@types";
 import { Loading } from "@components/Loading";
+import type { ButtonProps } from "./@types";
 
 export function Button(props: ButtonProps) {
   const {
-    type,
+    type = "button",
+    layout,
     className,
-    loadingClassName,
+    loadingClassName = "w-5",
     children,
     loading,
     disabled,
@@ -19,11 +20,12 @@ export function Button(props: ButtonProps) {
     className,
     loading && styles.loading,
     disabled && styles.disabled,
+    layout && styles[layout],
   );
 
   return (
-    <button onClick={onClick} type={type || "button"} className={classes}>
-      {loading ? <Loading className={loadingClassName || "w-5"} /> : children}
+    <button onClick={onClick} type={type} className={classes}>
+      {loading ? <Loading className={loadingClassName} /> : children}
     </button>
   );
 }