+++ src/components/Navbar/index.tsx
@@ -1,13 +1,8 @@
-import {
-  BellIcon,
-  GlobeIcon,
-  PlusIcon,
-  UserCircleIcon,
-  ViewGridIcon,
-} from "@heroicons/react/solid";
 import Link from "next/link";
 import { v4 as uuid } from "uuid";
 import styles from "./Navbar.module.css";
+import { Avatar } from "@components/Avatar";
+import { Icon } from "@components/Icon";
 import { useAccountContext } from "@contexts/AccountContext";
 import { PageUrls } from "@enums/pages";
 import { useClassName } from "@hooks/useClassName";
@@ -18,23 +13,23 @@ export function Navbar() {
   const { user } = useAccountContext();
   const items: NavigationItem[] = [
     {
-      icon: <BellIcon />,
+      icon: <Icon name="bell" type="fill" />,
       href: PageUrls.NOTIFICATIONS,
     },
     {
-      icon: <GlobeIcon />,
+      icon: <Icon name="globe" type="fill" />,
       href: PageUrls.SEARCH,
     },
     {
-      icon: <PlusIcon />,
+      icon: <Icon name="plus" type="fill" />,
       href: PageUrls.ADD_POST,
     },
     {
-      icon: <ViewGridIcon />,
+      icon: <Icon name="viewGrid" type="fill" />,
       href: PageUrls.HOME,
     },
     {
-      icon: <UserCircleIcon />,
+      icon: <Avatar size="tiny" />,
       href: `/${user?.username}`,
     },
   ];