+++ src/features/EditProfile/hooks/index.ts
@@ -1,5 +1,4 @@
 import { useEditProfile } from "./useEditProfile";
-import { useEditUser } from "./useEditUser";
 import { useEditUserValidations } from "./useEditUserValidations";
 
-export { useEditProfile, useEditUser, useEditUserValidations };
+export { useEditProfile, useEditUserValidations };