+++ src/features/Profile/components/Count/index.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import styles from "../Profile.module.css";
+import styles from "../../Profile.module.css";
 
 type Props = {
   className?: string;