+++ src/features/Profile/components/NotFound/index.tsx
@@ -1,5 +1,5 @@
-import { XIcon } from "@heroicons/react/outline";
-import styles from "../Profile.module.css";
+import styles from "../../Profile.module.css";
+import { Icon } from "@components/Icon";
 import { PageStatus } from "@components/PageStatus";
 
 export function NotFound() {
@@ -7,7 +7,7 @@ export function NotFound() {
     <div className={styles.notFound}>
       <PageStatus
         title="User Not Found"
-        icon={<XIcon strokeWidth="1" className="w-12" />}
+        icon={<Icon name="x" className="w-12" />}
       />
     </div>
   );