+++ src/features/Settings/@types/index.ts
@@ -11,3 +11,23 @@ export type SettingsItem = {
 export type LogoutButtonProps = {
   className?: string;
 };
+
+export type AvatarProfileProps = {
+  avatar?: string;
+};
+
+export type BackgroundProfileProps = {
+  background?: string;
+};
+
+export type ChangeImageProfileProps = {
+  title: string;
+  isOpen: boolean;
+  closeModal: () => void;
+  removeImage: () => void;
+  onClickInputFile: () => void;
+};
+
+export type UseImageProfileProps = {
+  key: "avatar" | "background";
+};