+++ src/graphql/@types/uploadFile.ts
@@ -1,0 +1,5 @@
+export type SingleUploadFileMutation = {
+  singleUploadFile: {
+    name: string;
+  };
+};