+++ src/hooks/useUrls.ts
@@ -1,0 +1,7 @@
+export function useUrls() {
+  function getFileUrl(fileName: string) {
+    return `${process.env.NEXT_PUBLIC_FILES_URL}/${fileName}`;
+  }
+
+  return { getFileUrl };
+}